refactor: reorganize new screening page files to fit desired structure
