+++ frontend/src/components/newScreening/ScreeningBreadcrumbs.tsx
@@ -48,7 +48,7 @@ function ScreeningBreadcrumbs({
             <BreadcrumbPage className="font-medium text-slate-700">Resultater</BreadcrumbPage>
           ) : canGoToResults ? (
             <BreadcrumbLink asChild className="hover:text-slate-700">
-              <Link to="/new-screening/results">Resultater</Link>
+              <Link to="/screening-historikk">Resultater</Link>
             </BreadcrumbLink>
           ) : (
             <BreadcrumbPage className="text-slate-400">Resultater</BreadcrumbPage>
