--- frontend/src/components/newScreening/newScreeningLib/screeningMockData.ts
@@ -1,152 +1,0 @@
-import type { ScreeningCandidate } from "@/components/newScreening/newScreeningLib/types";
-
-export const DEFAULT_JOB_FILE = "Fundamentals_of_Information_Systems_fullv1-komprimert.pdf";
-
-export const REQUIRED_SKILLS = [
-  "Python",
-  "React",
-  "Skyarkitektur",
-  "5+ års erfaring",
-  "Azure-sertifisering",
-];
-
-export const SCREENING_CANDIDATES: ScreeningCandidate[] = [
-  {
-    id: 1,
-    rank: 1,
-    name: "Lars Hansen",
-    role: "Senior programvareingeniør",
-    score: 92,
-    met: ["Python", "React", "Skyarkitektur", "5+ års erfaring"],
-    missing: ["Azure-sertifisering"],
-    summary:
-      "Erfaren programvareingeniør med 5+ års erfaring innen fullstack-utvikling, skyteknologi og smidig leveranse.",
-    experience: [
-      "Senior utvikler - Tech Company AS (2020-dd.)",
-      "Ledet utvikling av skybaserte applikasjoner",
-      "Implementerte CI/CD-pipelines",
-      "Veiledet juniorutviklere",
-      "Programvareingeniør - Digital Solutions (2018-2020)",
-      "Utviklet webapplikasjoner med moderne rammeverk",
-      "Samarbeidet med tverrfaglige team",
-    ],
-    education: [
-      "Master i informatikk",
-      "Norges teknisk-naturvitenskapelige universitet (NTNU)",
-    ],
-    email: "lars.hansen@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 2,
-    rank: 2,
-    name: "Emma Johansen",
-    role: "Fullstack-ingeniør",
-    score: 88,
-    met: ["JavaScript", "Node.js", "Teamledelse", "Smidig utvikling"],
-    missing: ["Docker", "Kubernetes"],
-    summary: "Sterk leveransehistorikk med JavaScript-økosystemer og teamledelse.",
-    experience: ["8 år i produkt- og plattformteam", "Ledet migrering til Node.js-mikrotjenester"],
-    education: ["Bachelor i programvareutvikling"],
-    email: "emma.j@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 3,
-    rank: 3,
-    name: "Ole Andersen",
-    role: "Backend-ingeniør",
-    score: 85,
-    met: ["Java", "Spring Boot", "Mikrotjenester", "CI/CD"],
-    missing: ["Frontend-erfaring", "React"],
-    summary: "Backend-orientert ingeniør med solid erfaring i mikrotjenestearkitektur.",
-    experience: ["7 år med backend-utvikling", "Bygget API-er med høy kapasitet"],
-    education: ["Master i informasjonssystemer"],
-    email: "ole.a@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 4,
-    rank: 4,
-    name: "Ingrid Larsen",
-    role: "Frontend-ingeniør",
-    score: 82,
-    met: ["TypeScript", "Angular", "Testarbeid", "Git"],
-    missing: ["Backend-erfaring", "Databasedesign"],
-    summary: "Frontend-spesialist med sterk TypeScript- og testpraksis.",
-    experience: ["6 år med frontend-leveranser", "Bygget applikasjoner basert på designsystem"],
-    education: ["Bachelor i informatikk"],
-    email: "ingrid.l@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 5,
-    rank: 5,
-    name: "Kristian Berg",
-    role: "Plattformingeniør",
-    score: 79,
-    met: ["C#", ".NET Core", "SQL Server"],
-    missing: ["Skyplattformer", "DevOps-erfaring"],
-    summary: "Solid .NET-profil med sterke SQL-grunnferdigheter.",
-    experience: ["5 år med plattformarbeid i større virksomhetsmiljøer"],
-    education: ["Bachelor i informasjonsteknologi"],
-    email: "kristian.b@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 6,
-    rank: 6,
-    name: "Sofie Nielsen",
-    role: "Programvareingeniør",
-    score: 76,
-    met: ["PHP", "Laravel", "MySQL"],
-    missing: ["Moderne frontend", "CI/CD", "Sky"],
-    summary: "Pålitelig backend-ingeniør med produksjonserfaring fra Laravel.",
-    experience: ["5 år med leveranse av interne forretningssystemer"],
-    education: ["Bachelor i dataingeniørfag"],
-    email: "sofie.n@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 7,
-    rank: 7,
-    name: "Magnus Olsen",
-    role: "Dataingeniør",
-    score: 73,
-    met: ["Python", "Django", "PostgreSQL"],
-    missing: ["React", "Skyplattformer", "Mikrotjenester"],
-    summary: "Data- og backend-orientert ingeniør med sterke Python-ferdigheter.",
-    experience: ["4 år med utvikling av datatunge applikasjoner"],
-    education: ["Master i datavitenskap"],
-    email: "magnus.o@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 8,
-    rank: 8,
-    name: "Anna Pedersen",
-    role: "Frontend-utvikler",
-    score: 70,
-    met: ["JavaScript", "Vue.js", "HTML/CSS"],
-    missing: ["Backend", "Databaser", "Testarbeid"],
-    summary: "Frontend-utvikler med sterke grunnferdigheter i UI-implementering.",
-    experience: ["4 år i kundevendte produktteam"],
-    education: ["Bachelor i multimedieteknologi"],
-    email: "anna.p@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-  {
-    id: 9,
-    rank: 9,
-    name: "Jonas Kristensen",
-    role: "Programvarekonsulent",
-    score: 68,
-    met: ["Ruby", "Rails", "Git"],
-    missing: ["JavaScript-rammeverk", "Sky", "Docker"],
-    summary: "Allsidig programvarekonsulent med erfaring i Rails.",
-    experience: ["6 år med konsulentarbeid på tvers av flere domener"],
-    education: ["Bachelor i programvareutvikling"],
-    email: "jonas.k@email.com",
-    phone: "+47 XXX XX XXX",
-  },
-];
