+++ frontend/src/main.tsx
@@ -1,14 +1,11 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter, Navigate, Route, Routes } from 'react-router'
+import { BrowserRouter, Route, Routes } from 'react-router'
 import '@/lib/styles/globals.css'
 import Home from '@/pages/Home'
 import Layout from '@/pages/Layout'
 import GeminiTestPage from '@/pages/GeminiTestPage'
-import ScreeningFlowPage from '@/pages/newScreening/ScreeningFlowPage'
-import NewScreeningPage from '@/pages/newScreening/NewScreeningPage'
-import ScreeningResultsPage from '@/pages/newScreening/ScreeningResultsPage'
-import ScreeningCandidatePage from '@/pages/newScreening/ScreeningCandidatePage'
+import NewScreeningPage from '@/pages/NewScreeningPage'
 import Screening from '@/pages/Screening'
 import ScreeningHistory from '@/pages/ScreeningHistory'
 import CVDatabase from '@/pages/CVDatabase'
@@ -21,11 +18,7 @@ createRoot(document.getElementById('root')!).render(
             <Route index element={<Home />} />
             <Route  path="GeminiTestPage" element={<GeminiTestPage />}/>
 
-            <Route path="/new-screening" element={<ScreeningFlowPage />}>
-              <Route index element={<NewScreeningPage />} />
-              <Route path="results" element={<ScreeningResultsPage />} />
-              <Route path="results/:candidateId" element={<ScreeningCandidatePage />} />
-            </Route>
+            <Route path="/new-screening" element={<NewScreeningPage />} />
 
             <Route path="/screening" element={<Screening />} />
             <Route path="/screening-historikk" element={<ScreeningHistory />} />
